Lighting Scene
Added a generic and simple test lighting scene for mason to try on 3D models. Its found in MOR_Scene folder called' MOR_LightingScene

Created a simple script that 'ping pongs' the light, its rudimentry but will do for now to see how 'flashing lights' affects objects. Change the speed in the inspector to change the rate of flashing